pyproject.toml
[project]
name = "shadow-optic"
version = "0.1.0"
description = "Autonomous Cost-Quality Optimization Engine for LLMs"
authors = [
    {name = "Shadow-Optic Team", email = "team@shadow-optic.dev"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["llm", "optimization", "portkey", "cost-reduction", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core - Portkey AI Gateway
    "portkey-ai>=2.0.0",
    
    # Workflow orchestration
    "temporalio>=1.7.0",
    
    # Evaluation
    "deepeval>=1.0.0",
    
    # Vector database
    "qdrant-client>=1.12.0",
    
    # ML & Data Science
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "joblib>=1.4.0",
    
    # P2 Enhancements - Minimal FOSS (Portkey-First)
    "tiktoken>=0.8.0",                 # Pre-request token estimation
    "dspy-ai>=2.5.0",                  # Via Portkey's native DSPy integration  
    "spotify-confidence>=2.0.0",       # Statistical significance testing
    
    # API Framework
    "pydantic>=2.10.0",
    "aiohttp>=3.10.0",
    "httpx>=0.28.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    
    # Observability
    "arize-phoenix>=4.0.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "structlog>=24.4.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.21.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
]

[project.urls]
Homepage = "https://github.com/shadow-optic/shadow-optic"
Documentation = "https://shadow-optic.dev/docs"
Repository = "https://github.com/shadow-optic/shadow-optic"
Issues = "https://github.com/shadow-optic/shadow-optic/issues"

[project.scripts]
shadow-optic = "shadow_optic.cli:main"
shadow-worker = "shadow_optic.worker:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shadow_optic"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
]

# ============================================================================
# Tool Configuration
# ============================================================================

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # `zip()` without an explicit `strict=` parameter
]

[tool.ruff.lint.isort]
known-first-party = ["shadow_optic"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "deepeval.*",
    "temporalio.*",
    "qdrant_client.*",
    "arize_phoenix.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=shadow_optic --cov-report=term-missing"

[tool.coverage.run]
source = ["src/shadow_optic"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
